requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
joblib>=1.3
pyarrow>=14.0
lightgbm>=4.0
optuna>=3.4
tqdm>=4.66
requests>=2.31
Pillow>=10.0
torch>=2.1
transformers>=4.41
sentencepiece>=0.1.99
peft>=0.11
trl>=0.9
accelerate>=0.34
unsloth>=2024.7.0
webdataset>=0.2.86
datasets>=2.20
